requirements.txt
discord.py==1.2.5
psycopg2==2.8.4
SQLAlchemy==1.3.12
